refactor: remove old user API routes and implement new user management endpoints
